fix: change project structure and separate listeners
