+++ pages/main/index.html
@@ -1,5 +1,5 @@
 <!DOCTYPE html>
-<html lang="en">
+<html lang="en" xmlns="http://www.w3.org/1999/html">
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
@@ -300,6 +300,7 @@
         </div>
     </div>
 </div>
-<script type="module" src="script.js" ></script>
+<script src="../../common/script.js"></script>
+<script src="script.js"></script>
 </body>
 </html>
