+++ pages/pets/index.html
@@ -32,6 +32,7 @@
                     <h3>Our friends who<br> are looking for a house</h3>
                     <section class="pets__cards-column">
                         <div class="pets__container-cards"></div>
+
                         <article class="pets__pagination">
                             <button  class="button_arrow button">
                                  <<
@@ -106,7 +107,7 @@
         </div>
     </div>
 </div>
-
-<script type="module" src="script.js" ></script>
+<script src="../../common/script.js"></script>
+<script  src="script.js" ></script>
 </body>
 </html>
