+++ pages/pets/script.js
@@ -1,3 +1,6 @@
-    import {pets, getCurrentPet,createCards } from "../main/script";
-    console.log(pets);
+const parentElem = document.querySelector('.pets__container-cards');
+
+let firstCard = pets.slice(0,8);
+
+createCards(firstCard,parentElem);
